+++ src/embed_colors.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-// figure out how to do geode colors next update
 namespace embed_color {
 
 inline constexpr int kGameClose = 0x992D22;